feat: add functionality to display companies from store
