+++ package.json
@@ -17,9 +17,13 @@
   },
   "dependencies": {
     "next": "10.2.2",
+    "next-redux-wrapper": "^6.0.2",
     "prop-types": "^15.7.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-redux": "^7.2.4",
+    "redux": "^4.1.0",
+    "redux-thunk": "^2.3.0",
     "uuid": "^8.3.2",
     "uuidv4": "^6.2.8"
   },
