+++ pages/companies/index.js
@@ -1,105 +1,27 @@
-import PropTypes from 'prop-types';
+// import PropTypes from 'prop-types';
+import { useSelector } from 'react-redux';
 import Head from 'next/head';
 import Link from 'next/link';
-import { useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
+// import { useEffect } from 'react';
 import styles from '../../styles/Companies.module.css';
-import server from '../../config';
-import Modal from '../../components/Modal';
+import AddBook from '../../components/addCompany';
 
-export const getStaticProps = async () => {
-  const response = await fetch(`${server}/api/companies`);
-  const companiesInArray = await response.json();
+const Companies = () => {
+  const companies = useSelector(state => state.companies);
+  // const dispatch = useDispatch();
 
-  return {
-    props: { companiesArray: companiesInArray },
-  };
-};
-
-const Companies = ({ companiesArray }) => {
-  const [values, setValues] = useState({
-    id: '',
-    name: '',
-    ceo: '',
-    about: '',
-  });
-  const [companies, setCompanies] = useState(companiesArray);
-  const handleSubmit = e => {
-    e.preventDefault();
-    setCompanies([values, ...companies]);
-    setValues({
-      ...values,
-      id: '',
-      name: '',
-      ceo: '',
-      about: '',
-    });
-  };
+  // useEffect(() => {
+  //   dispatch(fetchCompany(companies));
+  // }, [companies]);
   const handleDelete = id => {
-    const updatedCompanies = companies.filter(company => company.id !== id);
-    setCompanies(updatedCompanies);
+    console.log(id);
   };
   return (
     <>
       <Head>
         <title>Companies List | Lists</title>
       </Head>
-      <div className="p-10 mt-10 mb-10">
-        <div>
-          <button
-            type="button"
-            data-target="add-company-modal"
-            className="open-modal-button btn btn-primary"
-          >
-            Add Company
-          </button>
-          <Modal modalId="add-company-modal" modalTitle="Add Company">
-            <form onSubmit={handleSubmit}>
-              <div className="form-group titleInputDiv mb-10">
-                <input
-                  type="text"
-                  className="form-control w-full"
-                  id="company-name"
-                  name="name"
-                  placeholder="Enter Company Name"
-                  value={values.name}
-                  onChange={e => setValues({ ...values, id: uuidv4(), name: e.target.value })}
-                  required
-                />
-              </div>
-              <div className="form-group titleInputDiv mb-10">
-                <input
-                  type="text"
-                  className="form-control w-full"
-                  id="company-ceo-name"
-                  name="ceo"
-                  placeholder="Enter Company CEO Name"
-                  value={values.ceo}
-                  onChange={e => setValues({ ...values, id: uuidv4(), ceo: e.target.value })}
-                  required
-                />
-              </div>
-              <div className="form-group titleInputDiv mb-10">
-                <textarea
-                  className="form-control w-full"
-                  id="about-company"
-                  name="about"
-                  rows="3"
-                  placeholder="About Company"
-                  value={values.about}
-                  onChange={e => setValues({ ...values, id: uuidv4(), about: e.target.value })}
-                  required
-                />
-              </div>
-              <div className="my-10">
-                <button type="submit" className="btn d-block btn-primary w-full mx-auto">
-                  Submit
-                </button>
-              </div>
-            </form>
-          </Modal>
-        </div>
-      </div>
+      <AddBook />
       <h1 className="ta-center">All Companies</h1>
       {companies.length >= 1 ? (
         companies.map(company => (
@@ -128,10 +50,6 @@ const Companies = ({ companiesArray }) => {
       )}
     </>
   );
-};
-
-Companies.propTypes = {
-  companiesArray: PropTypes.arrayOf(PropTypes.object).isRequired,
 };
 
 export default Companies;
